refactor: Consolidate YAML tests and clean up code
- Moved regression tests to existing yaml.test.ts file
- Removed separate test files for better organization
- Removed unnecessary code comments while keeping functionality
- All tests pass with fixes and fail on stable (verified)
